[fs] Fixed invalid FSstat object when file is deleted
We are not currently checking if the file exists when
calling FS.statSync() to create the Stats object, in which
case if the file doesn't exists, we should return a undefined
value, and the unit tests should also check if the value
is undefined.  Added a ToDo to properly fix this with adding
a FS.exist() api that can be used to check if the file exists,
and then return a valid FSStat object when the file is invalid
according the NodeJS spec.

Also updated tests and API doc to reflect new change.

Fixes #1205

+++ tests/test-fs.js
@@ -5,7 +5,7 @@ var assert = require("Assert.js");
 
 // Clean up any test files left from previous tests
 var stats = fs.statSync('testfile.txt');
-if (stats.isFile() || stats.isDirectory()) {
+if (stats && (stats.isFile() || stats.isDirectory())) {
     fs.unlinkSync('testfile.txt');
     console.log('removing testfile.txt');
 }
@@ -15,7 +15,7 @@ var fd = fs.openSync('testfile.txt', 'a');
 fs.closeSync(fd);
 var fd_stats = fs.statSync('testfile.txt');
 var success = false;
-if (fd_stats.isFile()) {
+if (fd_stats && fd_stats.isFile()) {
 	success = true;
 }
 assert(success, "create file");
@@ -24,7 +24,7 @@ assert(success, "create file");
 fs.unlinkSync('testfile.txt');
 fd_stats = fs.statSync('testfile.txt');
 success = false;
-if (!fd_stats.isFile()) {
+if (!fd_stats || !fd_stats.isFile()) {
 	success = true;
 }
 assert(success, "remove file");
